ShelfLifeDAM-main/backend/requirements.txt
# Core Django
Django==4.2.7
djangorestframework==3.16.1
djangorestframework-simplejwt==5.3.0

# Wagtail CMS
wagtail==7.1.1
django-modelcluster==6.4
django-taggit==6.1.0
django-treebeard==4.7.1
django-permissionedforms==0.1
draftjs_exporter==5.1.0
Willow==1.11.0
pillow-heif==1.1.1
laces==0.1.2
django-tasks==0.8.1
telepath==0.3.1

# API and CORS
django-cors-headers==4.3.1
django-filter==23.5
drf-yasg==1.21.7

# Database
psycopg2-binary==2.9.9

# Authentication
PyJWT==2.10.1

# File Storage
django-storages==1.14.2
boto3==1.29.7
botocore==1.32.7
s3transfer==0.8.0
Pillow==11.3.0

# Task Queue (Celery)
celery==5.3.4
redis==5.0.1
kombu==5.5.4
billiard==4.2.2
amqp==5.3.1
vine==5.1.0

# Utilities
python-dotenv==1.0.0
python-dateutil==2.9.0.post0
pytz==2025.2
whitenoise==6.6.0
requests==2.32.5
beautifulsoup4==4.14.2
anyascii==0.3.3

# Dependencies
asgiref==3.10.0
certifi==2025.10.5
charset-normalizer==3.4.3
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
defusedxml==0.7.1
django-stubs-ext==5.2.7
et_xmlfile==2.0.0
filetype==1.2.0
idna==3.11
inflection==0.5.1
jmespath==1.0.1
openpyxl==3.1.5
packaging==25.0
prompt_toolkit==3.0.52
PyYAML==6.0.3
setuptools==80.9.0
six==1.17.0
soupsieve==2.8
sqlparse==0.5.3
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.0.7
wcwidth==0.2.14